agent/requirements.txt
fastapi
uvicorn
requests
pyyaml
prometheus-client
ansible-runner
